Assignments/assignment3/pyproject.toml
[project]
name = "assignment3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cryptography>=46.0.4",
    "paho-mqtt>=2.1.0",
    "pyqt5>=5.15.11",
]
